add two classes -> TransactionDAOTest and BankDAOTest, add the annotations to Transaction and Bank class @Accessors(chain = true)

+++ src/main/java/org/example/dao/BankDAO.java
@@ -32,7 +32,7 @@ public class BankDAO extends DataAccsessObject<Bank> {
             ResultSet resultSet = statement.executeQuery();
             while (resultSet.next()) {
 
-                bank.setId(resultSet.getInt("id"));
+                bank.setId(resultSet.getLong("id"));
                 bank.setName(resultSet.getString("name"));
 
             }
